qcacld-2.0: Fix a potential overflow issue
Overflow may happen in function ol_tx_vdev_set_bundle_require
while calculating vdev->bundling_reqired as both high_th and
low_th are configurable in range of {0, 70000}, which will
cause false triggering for bundle. To address this, change
variable high_th/low_th to uint64_t.

CRs-Fixed: 2816823

+++ drivers/net/wireless/qcacld-2.0/CORE/SERVICES/COMMON/ol_txrx_osif_api.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2014 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2012, 2014, 2017-2018, 2020 The Linux Foundation. All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -211,14 +211,14 @@ adf_nbuf_t ol_txrx_osif_tso_segment(
 #ifdef QCA_SUPPORT_TXRX_HL_BUNDLE
 void
 ol_tx_vdev_set_bundle_require(uint8_t vdev_id, unsigned long tx_packets,
-			uint32_t time_in_ms, uint32_t high_th, uint32_t low_th);
+			uint32_t time_in_ms, uint64_t high_th, uint64_t low_th);
 void
 ol_tx_pdev_reset_bundle_require(void* pdev);
 
 #else
 static inline void
 ol_tx_vdev_set_bundle_require(uint8_t vdev_id, unsigned long tx_packets,
-			uint32_t time_in_ms, uint32_t high_th, uint32_t low_th)
+			uint32_t time_in_ms, uint64_t high_th, uint64_t low_th)
 {
 	return;
 }
